Moved tests around to make them easier to coverage test
