+++ .vscode/settings.json
@@ -1,3 +1,6 @@
 {
-    "python.REPL.enableREPLSmartSend": false
+    "python.REPL.enableREPLSmartSend": false,
+    "python.testing.unittestEnabled": false,
+    "python.testing.pytestEnabled": true,
+    "python.testing.pytestPath": "/Users/borrow-adm/Documents/SimonsObs/librarian/env311/bin/pytest",
 }
